move : StationPair 를 vo 로 이동
