+++ src/main/java/subway/application/LineService.java
@@ -10,7 +10,7 @@ import subway.dao.StationDao;
 import subway.domain.Line;
 import subway.domain.Section;
 import subway.domain.Station;
-import subway.domain.StationPair;
+import subway.vo.StationPair;
 import subway.domain.Stations;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
