+++ src/main/java/subway/domain/Stations.java
@@ -10,6 +10,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import subway.exception.IllegalStationsException;
+import subway.vo.StationPair;
 
 public class Stations {
