+++ src/test/java/subway/dao/LineDaoTest.java
@@ -11,9 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Station;
-import subway.domain.StationPair;
+import subway.exception.IllegalStationsException;
+import subway.vo.StationPair;
 
 @DisplayName("라인 Dao 테스트")
 @Transactional
@@ -129,5 +128,15 @@ class LineDaoTest {
         assertThat(result)
             .isPresent()
             .hasValue(response);
+    }
+
+    private void validateDuplicateUpStation(List<StationPair> stationPairs) {
+        long distinctUpStationCount = stationPairs.stream()
+            .map(StationPair::getUpStation)
+            .distinct()
+            .count();
+        if (distinctUpStationCount != stationPairs.size()) {
+            throw new IllegalStationsException("중복된 역은 노선에 포함될 수 없습니다.");
+        }
     }
 }
