+++ src/test/java/subway/domain/StationPairTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import subway.exception.IllegalStationsException;
+import subway.vo.StationPair;
 
 @DisplayName("StationPair 테스트")
 class StationPairTest {
